requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
blockcypher==1.0.83
pyqrcode==1.2.1
